pyproject.toml
[project]
name = "pie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "adafruit-circuitpython-dht>=4.0.9",
    "gpiozero>=2.0.1",
    "lgpio>=0.2.2.0",
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "watchdog>=6.0.0",
]
